touchscreen: sun4i-ts: A10 (sun4i) has a different temperature curve
Testing has revealed that the temperature in the rtp controller of the A10
(sun4i) SoC has a different curve then on the A13 (sun5i) and later models.

Add a new sun5i-a13-ts compatible to differentiate the newer models and
set the curve based on the compatible string.

This fixes the temperature reported on the A10 being much higher then
expected.

Note the new curve is still not ideal on all A10-s, that seems to have to
do with there being a large spread between different A10-s out there.

Reported-by: Tong Zhang <lovewilliam@gmail.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>